+++ ui-web-react/src/components/races/race-list.container.ts
@@ -2,7 +2,7 @@ import { StoredState } from "../../model/types/datatypes";
 import RaceListComponent from "./race-list.component";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { racesRequested } from "../../model/actions/actions";
+import { racesRequested } from "../../model/actions/race.actions";
 
 const mapStateToProps = (state: StoredState) => {
   return {
