+++ ui-web-react/src/model/sagas/alerts.saga.ts
@@ -1,5 +1,5 @@
 import { put, takeEvery } from "redux-saga/effects";
-import { ALERTS_SHOW, AlertsAction, alertsHide } from "../actions/actions";
+import { ALERTS_SHOW, AlertsAction, alertsHide } from "../actions/alerts.actions";
 import { DEFAULT_ALERTS_TIMEOUT } from "../utils/constants";
 import { delay } from "./sagas";
